move vehicle data parsing to separate model file
